Some renaming to make component jobs more clear.
