+++ src/modules/blocks/Blocks.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 
-import NextBlock from './NextBlock.jsx'
+import BlockIterator from './BlockIterator.jsx'
 import Logger from 'logger';
 
 import { API_V1 } from 'http/url.js';
@@ -34,7 +34,7 @@ export default class Blocks extends Component {
   render() {
     return(
         <div className={styles.wrapper}>
-          <NextBlock
+          <BlockIterator
             triggerOut={this.props.triggerOut}
             triggerAction={this.props.actionOut}
             blocks={this.props.blocks || this.state.blocks} />
